requirements.txt
BeautifulSoup
Django == 1.4.5
MySQL-python
Pillow < 2.0
PyICU
PyYAML
django-piston
Pygments
RBTools
South
django-appconf
django-compressor
django-imagekit
git+git://github.com/GrandComicsDatabase/django-taggit.git
git+git://github.com/GrandComicsDatabase/django-templatesadmin.git
git+git://github.com/GrandComicsDatabase/django-mobile.git
python-stdnum
wsgiref
